chore: move const content into fund specific files + fix tests
