+++ catalyst_voices/packages/internal/catalyst_voices_models/lib/src/catalyst_voices_models.dart
@@ -30,7 +30,11 @@ export 'crypto/lock_factor.dart';
 export 'dev_tools/dev_tools_config.dart';
 export 'document/builder/document_builder.dart';
 export 'document/builder/document_change.dart';
-export 'document/constant_documents_refs.dart';
+export 'document/constant/active_const_documents_refs.dart';
+export 'document/constant/all_const_documents_refs.dart';
+export 'document/constant/constant_documents_refs.dart';
+export 'document/constant/f14_const_documents_refs.dart';
+export 'document/constant/f15_const_documents_refs.dart';
 export 'document/data/document_data.dart';
 export 'document/data/document_data_content.dart';
 export 'document/data/document_data_metadata.dart';
