+++ catalyst_voices/packages/internal/catalyst_voices_models/lib/src/document/constant/active_const_documents_refs.dart
@@ -1,0 +1,3 @@
+import 'package:catalyst_voices_models/catalyst_voices_models.dart';
+
+const List<CategoryTemplatesRefs> activeConstantDocumentRefs = f15ConstDocumentsRefs;
