+++ catalyst_voices/packages/internal/catalyst_voices_models/lib/src/document/constant/all_const_documents_refs.dart
@@ -1,0 +1,6 @@
+import 'package:catalyst_voices_models/catalyst_voices_models.dart';
+
+final List<CategoryTemplatesRefs> allConstantDocumentRefs = [
+  ...f14ConstDocumentsRefs,
+  ...f15ConstDocumentsRefs,
+];
