+++ catalyst_voices/packages/internal/catalyst_voices_services/lib/src/comment/comment_service.dart
@@ -46,7 +46,7 @@ final class CommentServiceImpl implements CommentService {
   Future<CommentTemplate> getCommentTemplateFor({
     required DocumentRef category,
   }) async {
-    final commentTemplateRef = constantDocumentsRefs
+    final commentTemplateRef = activeConstantDocumentRefs
         .firstWhereOrNull((element) => element.category.id == category.id)
         ?.comment;
