+++ catalyst_voices/pubspec.yaml
@@ -4,7 +4,7 @@ environment:
   sdk: ">=3.3.3 <4.0.0"
 
 dependencies:
-  dart_pubspec_licenses: ^3.0.4
+  dart_pubspec_licenses: 3.0.4
 
 dev_dependencies:
   melos: ^6.3.3
